Thống kê sản phẩm, hóa đơn

+++ src/main/java/com/example/datn/repository/HoaDonRepo.java
@@ -10,7 +10,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HoaDonRepo extends JpaRepository<HoaDon, Integer> {
+public interface HoaDonRepo extends JpaRepository<HoaDon, Long> {
     @Query("select hd from HoaDon hd order by hd.ngayTao desc ")
     Page<HoaDon> findHoaDonAndSortDay(Pageable pageable);
 
@@ -30,7 +30,7 @@ public interface HoaDonRepo extends JpaRepository<HoaDon, Integer> {
             "FROM HoaDon hd " +
             "join hd.phieuGiamGia pgg " +
             "where hd.id =:hoaDonId")
-    HoaDonResponse findPGGByHoaDonId(@Param("hoaDonId") int hoaDonId);
+    HoaDonResponse findPGGByHoaDonId(@Param("hoaDonId") long hoaDonId);
 
     Page<HoaDon> findAllByTrangThai(Integer trangThai, Pageable pageable);
