+++ src/main/java/com/example/datn/repository/LichSuHoaDonRepo.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface LichSuHoaDonRepo extends JpaRepository<LichSuHoaDon,Integer> {
+public interface LichSuHoaDonRepo extends JpaRepository<LichSuHoaDon,Long> {
     @Query("select l from LichSuHoaDon l where l.hoaDon.id=:idHoaDon")
-    List<LichSuHoaDon> findLichSuHoaDonByIdHoaDon(@Param("idHoaDon") int idHoaDon);
+    List<LichSuHoaDon> findLichSuHoaDonByIdHoaDon(@Param("idHoaDon") long idHoaDon);
 }
